make utxo and evm optional deps
