+++ .github/workflows/mysql_test.yml
@@ -46,7 +46,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install .[test]
+        pip install .[test, utxo]
     
     - name: Run tests and generate coverage
       run: |
