+++ .github/workflows/test.yml
@@ -38,7 +38,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install pytest pytest-cov
-        pip install .[test]
+        pip install .[test,evm,utxo]
 
     - name: Run tests and generate coverage
       run: |
