+++ aiotx/clients/_utxo_base_client.py
@@ -1,18 +1,11 @@
 import asyncio
 import json
+import sys
 from decimal import Decimal
 from typing import Optional, Union
 
 import aiohttp
 from aiohttp.client_exceptions import ClientError, ClientOSError
-from bitcoinlib.encoding import pubkeyhash_to_addr_bech32
-from bitcoinlib.keys import HDKey, Key
-from bitcoinlib.networks import Network
-from bitcoinlib.transactions import Transaction
-from sqlalchemy import BigInteger, Boolean, Column, Integer, String, select, update
-from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
-from sqlalchemy.orm import declarative_base, sessionmaker
-from sqlalchemy.pool import NullPool
 
 from aiotx.clients._base_client import AioTxClient, BlockMonitor
 from aiotx.exceptions import (
@@ -28,43 +21,6 @@ from aiotx.exceptions import (
 from aiotx.log import logger
 from aiotx.types import FeeEstimate
 
-Base = declarative_base()
-
-
-def create_address_model(currency_name):
-    class Address(Base):
-        __tablename__ = f"{currency_name}_addresses"
-        __table_args__ = {"extend_existing": True}
-
-        address = Column(String(255), primary_key=True)
-        block_number = Column(Integer)
-
-    return Address
-
-
-def create_utxo_model(currency_name):
-    class UTXO(Base):
-        __tablename__ = f"{currency_name}_utxo"
-        __table_args__ = {"extend_existing": True}
-
-        tx_id = Column(String(255), primary_key=True)
-        output_n = Column(Integer, primary_key=True)
-        address = Column(String(255))
-        amount_satoshi = Column(BigInteger)
-        used = Column(Boolean, default=False)
-
-    return UTXO
-
-
-def create_last_block_model(currency_name):
-    class LastBlock(Base):
-        __tablename__ = f"{currency_name}_last_block"
-        __table_args__ = {"extend_existing": True}
-
-        block_number = Column(Integer, primary_key=True)
-
-    return LastBlock
-
 
 class AioTxUTXOClient(AioTxClient):
     def __init__(
@@ -77,6 +33,17 @@ class AioTxUTXOClient(AioTxClient):
         network_name,
         db_url,
     ):
+        try:
+            from bitcoinlib.networks import Network
+
+        except ImportError:
+            logger.error(
+                "The required dependencies for AioTxLTCClient or AioTxLTCClient are not installed. "
+                "Please install the necessary packages and try again."
+                "pip install aiotx[utxo]"
+            )
+            sys.exit(-1)
+
         super().__init__(node_url, headers)
         self.node_username = node_username
         self.node_password = node_password
@@ -94,6 +61,9 @@ class AioTxUTXOClient(AioTxClient):
         return amount / 10**8
 
     async def generate_address(self) -> dict:
+        from bitcoinlib.encoding import pubkeyhash_to_addr_bech32
+        from bitcoinlib.keys import HDKey
+
         hdkey = HDKey()
         derivation_path = f"m/84'/{self._network.bip44_cointype}'/0'/0/0"
         private_key = hdkey.subkey_for_path(derivation_path).private_hex
@@ -106,6 +76,9 @@ class AioTxUTXOClient(AioTxClient):
         return private_key, address
 
     def get_address_from_private_key(self, private_key):
+        from bitcoinlib.encoding import pubkeyhash_to_addr_bech32
+        from bitcoinlib.keys import Key
+
         key = Key(private_key)
         address = pubkeyhash_to_addr_bech32(
             key.hash160, prefix=self._network.prefix_bech32, witver=0, separator="1"
@@ -266,7 +239,9 @@ class AioTxUTXOClient(AioTxClient):
         from_address: str,
         fee: int,
         deduct_fee: bool,
-    ) -> tuple[Transaction, list]:
+    ):
+        from bitcoinlib.transactions import Transaction
+
         transaction = Transaction(network=self._network.name, witness_type="segwit")
         inputs = []
         outputs = []
@@ -312,9 +287,9 @@ class AioTxUTXOClient(AioTxClient):
 
         return transaction, inputs, outputs
 
-    def _sign_transaction(
-        self, transaction: Transaction, private_keys: list[str]
-    ) -> Transaction:
+    def _sign_transaction(self, transaction, private_keys: list[str]):
+        from bitcoinlib.keys import Key
+
         for i, private_key in enumerate(private_keys):
             key = Key(private_key)
             transaction.sign(key, i)
@@ -378,6 +353,16 @@ class AioTxUTXOClient(AioTxClient):
 
 class UTXOMonitor(BlockMonitor):
     def __init__(self, client: AioTxUTXOClient, db_url):
+        from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
+        from sqlalchemy.orm import sessionmaker
+        from sqlalchemy.pool import NullPool
+
+        from aiotx.utils.utxo_db_models import (
+            create_address_model,
+            create_last_block_model,
+            create_utxo_model,
+        )
+
         self.client = client
         self.block_handlers = []
         self.transaction_handlers = []
@@ -450,6 +435,8 @@ class UTXOMonitor(BlockMonitor):
                 await handler(transaction)
 
     async def _init_db(self):
+        from aiotx.utils.utxo_db_models import Base
+
         async with self._engine.begin() as conn:
             await conn.run_sync(Base.metadata.create_all)
             try:
@@ -462,10 +449,13 @@ class UTXOMonitor(BlockMonitor):
                 await self._init_last_block(last_known_block)
 
     async def _drop_tables(self):
+        from aiotx.utils.utxo_db_models import Base
+
         async with self._engine.begin() as conn:
             await conn.run_sync(Base.metadata.drop_all)
 
     async def _add_new_address(self, address: str, block_number: int = None):
+        from sqlalchemy import select
         last_known_block = await self.client.get_last_block_number()
         if block_number is None:
             block_number = last_known_block
@@ -489,6 +479,8 @@ class UTXOMonitor(BlockMonitor):
     async def _add_new_utxo(
         self, address: str, tx_id: str, amount: int, output_n: int
     ) -> None:
+        from sqlalchemy import select
+
         async with self._session() as session:
             async with session.begin():
                 existing_utxo = await session.scalar(
@@ -510,6 +502,7 @@ class UTXOMonitor(BlockMonitor):
                 await session.commit()
 
     async def _update_last_block(self, block_number: int) -> None:
+        from sqlalchemy import select
         async with self._session() as session:
             async with session.begin():
                 data = await session.execute(select(self.LastBlock))
@@ -521,6 +514,8 @@ class UTXOMonitor(BlockMonitor):
                     await session.commit()
 
     async def _init_last_block(self, block_number: int) -> None:
+        from sqlalchemy import select
+
         async with self._session() as session:
             async with session.begin():
                 data = await session.execute(select(self.LastBlock))
@@ -535,6 +530,8 @@ class UTXOMonitor(BlockMonitor):
             await self._delete_utxo(txid, vout)
 
     async def _get_utxo_data(self, address: str, spent=False):
+        from sqlalchemy import select
+
         async with self._session() as session:
             result = await session.execute(
                 select(
@@ -547,6 +544,8 @@ class UTXOMonitor(BlockMonitor):
             return result.fetchall()
 
     async def _mark_utxo_used(self, tx_id: str, output_n: str) -> None:
+        from sqlalchemy import update
+
         async with self._session() as session:
             async with session.begin():
                 stmt = (
@@ -560,6 +559,8 @@ class UTXOMonitor(BlockMonitor):
                 await session.commit()
 
     async def _get_utxo(self, tx_id: str, output_n: int):
+        from sqlalchemy import select
+
         async with self._session() as session:
             result = await session.execute(
                 select(self.UTXO).where(
@@ -569,6 +570,8 @@ class UTXOMonitor(BlockMonitor):
             return result.scalar()
 
     async def _get_all_utxo_tx_ids(self):
+        from sqlalchemy import select
+
         """
         Function to optimize UTXO checking on new block to not make a SELECT for each input in the block
         """
@@ -583,11 +586,15 @@ class UTXOMonitor(BlockMonitor):
                 await session.commit()
 
     async def _get_last_block(self) -> Optional[int]:
+        from sqlalchemy import select
+
         async with self._session() as session:
             result = await session.execute(select(self.LastBlock.block_number))
             return result.scalar()
 
     async def _get_addresses(self) -> list[str]:
+        from sqlalchemy import select
+
         async with self._session() as session:
             result = await session.execute(select(self.Address.address))
             return {row[0] for row in result.fetchall()}
